+++ browser/net/decentralized_dns_network_delegate_helper.cc
@@ -20,63 +20,7 @@
 #include "content/public/browser/browser_context.h"
 #include "net/base/net_errors.h"
 
-namespace {
-// Ipfs codes from multicodec table
-// https://github.com/multiformats/multicodec/blob/master/table.csv
-const int64_t kIpfsNSCodec = 0xE3;
-const int64_t kIpnsNSCodec = 0xE5;
-inline constexpr char kIPFSScheme[] = "ipfs";
-inline constexpr char kIPNSScheme[] = "ipns";
-
-// Decodes a varint from the given string piece into the given int64_t. Returns
-// remaining span if the string had a valid varint (where a byte was found with
-// it's top bit set).
-base::span<const uint8_t> DecodeVarInt(base::span<const uint8_t> from,
-                                       int64_t* into) {
-  auto it = from.begin();
-  int shift = 0;
-  uint64_t ret = 0;
-  do {
-    if (it == from.end()) {
-      return {};
-    }
-
-    // Shifting 64 or more bits is undefined behavior.
-    DCHECK_LT(shift, 64);
-    unsigned char c = *it;
-    ret |= static_cast<uint64_t>(c & 0x7f) << shift;
-    shift += 7;
-  } while (*it++ & 0x80);
-  *into = static_cast<int64_t>(ret);
-  return from.subspan(it - from.begin());
-}
-
-}  // namespace
-
 namespace decentralized_dns {
-
-GURL ContentHashToCIDv1URL(base::span<const uint8_t> contenthash) {
-  int64_t code = 0;
-  contenthash = DecodeVarInt(contenthash, &code);
-  if (contenthash.empty()) {
-    return GURL();
-  }
-  if (code != kIpnsNSCodec && code != kIpfsNSCodec) {
-    return GURL();
-  }
-  std::string encoded = base32::Base32Encode(contenthash);
-  if (encoded.empty()) {
-    return GURL();
-  }
-  std::string trimmed;
-  base::TrimString(encoded, "=", &trimmed);
-  std::string lowercase = base::ToLowerASCII(trimmed);
-  // multibase format <base-encoding-character><base-encoded-data>
-  // https://github.com/multiformats/multibase/blob/master/multibase.csv
-  std::string cidv1 = "b" + lowercase;
-  std::string scheme = (code == kIpnsNSCodec) ? kIPNSScheme : kIPFSScheme;
-  return GURL(scheme + "://" + cidv1);
-}
 
 int OnBeforeURLRequest_DecentralizedDnsPreRedirectWork(
     const brave::ResponseCallback& next_callback,
