+++ components/brave_wallet/browser/json_rpc_service_unittest.cc
@@ -31,7 +31,7 @@
 #include "base/test/scoped_feature_list.h"
 #include "base/test/task_environment.h"
 #include "base/values.h"
-#include "brave/browser/net/decentralized_dns_network_delegate_helper.h"
+#include "brave/components/decentralized_dns/core/utils.h"
 #include "brave/components/brave_wallet/browser/blockchain_list_parser.h"
 #include "brave/components/brave_wallet/browser/blockchain_registry.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_constants.h"
